requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anthropic==0.52.2
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.32
botocore==1.38.32
Brotli==1.1.0
browser-use==0.2.5
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
configparser==7.2.0
cssselect2==0.8.0
defusedxml==0.7.1
distro==1.9.0
docx2pdf==0.1.8
faiss-cpu==1.11.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.2
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.31.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
grpcio==1.72.1
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.4
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.45
lxml==5.4.0
Mako==1.3.10
Markdown==3.8
markdownify==1.1.0
MarkupSafe==3.0.2
mem0ai==0.1.93
monotonic==1.6
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
ollama==0.5.1
openai==1.101.0
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
patchright==1.52.5
pillow==11.2.1
playwright==1.52.0
portalocker==2.10.1
posthog==3.25.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.11.0
pyee==13.0.0
pyparsing==3.2.3
pyperclip==1.9.0
pyphen==0.17.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2024.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
screeninfo==0.8.1
selenium==4.33.0
sentence-transformers==4.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
undetected-chromedriver==3.5.5
uritemplate==4.2.0
urllib3==2.4.0
uuid7==0.1.0
weasyprint==65.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wordlist==1.0.1
wsproto==1.2.0
xlsxwriter==3.2.5
zopfli==0.2.3.post1
zstandard==0.23.0
